tdf#123912 ww8 export: re-protect implicit section
Sections should retain their protected on/off status
regardless of the value of PROTECT_FORM. However, if there
ARE no sections, then the implicit section should use the
document settings. The same is true for the pseudo -1 section
which I believe can only be the last section (the fragment
of the implicit section that follows the last real section).

This is basically a revert of LO 6.2 commit
fa667b6dc410f3af57ef436cc117352c829f95e7, restoring
the previous behaviour in the case of the implicit section.

Change-Id: If0b473445e0add017504a3cb61b63116f92be5ce
Reviewed-on: https://gerrit.libreoffice.org/69957
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>