tdf#124261 sw btlr writing mode render: fix invalidation rectangle
SwTextFrame::SwitchHorizontalToVertical(SwRect) did not handle the non-swapped
case, also the text frame took the tblr codepath when determining the
invalidation rectangle of the layout frame, but that special-case is not
needed for the btlr direction.

Typing into the cell of the bugdoc now determines the paint rectangle
correctly, so the typed letters show up right after typing.

Change-Id: I48c75603963868591d72bf4d0e8dbba679f3b5e9
Reviewed-on: https://gerrit.libreoffice.org/70087
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins