tdf#114316 vcl opengl windows: fix missing context menu in full-screen mode
Full-screen mode on Windows used to work by measuring the space needed
by window caption (title) and borders, then positioning and sizing the
window in a way, so that the caption and borders are not visible.

This approach breaks at least in the OpenGL case where a large enough
negative position results in rendering errors.

Fix the problem by explicitly requesting the window to have no caption,
so we render less outside the screen (30 pixels -> 8 pixels in my case),
which makes the "exit fullscreen" toolbar appear, also the context menu
is visible.

Change-Id: I6cf2b9774b505d3887b958a6a018b5ae84bbe4bc
Reviewed-on: https://gerrit.libreoffice.org/70191
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins