sw btlr writing mode: implement RTF export
Note that we can't work with the table box (and frame format) of
pTableTextNodeInfoInner in RtfAttributeOutput::TableVerticalCell(). This
is because cell definitions are written in
RtfAttributeOutput::TableDefinition(), and the loop only calls
setCell(), i.e. the cell index is updated in the node info, but the
table box is not.

Import was already working as-is.

Change-Id: Ia089c42b3a49f1a8f012a8cadaa8bcad16128c3e
Reviewed-on: https://gerrit.libreoffice.org/70269
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>