tdf#124944 don't directly adjust fallback layouts
These are adjusted via MultiSalLayout::AdjustLayout, as they
just know their own text width, but not the full text width of
the MultiSalLayout, so the adjustment is too large.

The correct adjustment is done by preparing a DXArray for all
glyphs, which is then passed to the individual font fallback
layouts for general adjustment.

Change-Id: I439527d03df4752e74ff94aee7775a95275be5e2
Reviewed-on: https://gerrit.libreoffice.org/71440
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>