sw/qa: recover locale in testUserFieldTypeLanguage.
In testUserFieldTypeLanguage the locale was set to de-DE
first then revert by calling SetLocaleConfigString with an
empty string when the test finished. Calling
SetLocaleConfigString with an empty string set the system
locale to the system default, which is not necessarily en-US
which is the default for unit tests.

Unfortunately it makes testTdf108021 fail because the
layouts in the testTdf108021 are different in en-US and
zh-TW locale in my environment. This locale-dependent layout
issue is tracked as tdf#125025 now.

Change-Id: I50ae2df5fb4624f94d070742c6189bf333370137
Reviewed-on: https://gerrit.libreoffice.org/71480
Tested-by: Jenkins
Reviewed-by: Mark Hung <marklh9@gmail.com>