tdf#125151 sw: allow protection around field marks
Similar to bug 105625, the cursor should be considered outside
of a fieldmark when it is at the start. A delete at the start would
normally delete the entire form field, but in this case the section
is supposed to be protected, so deleting should be forbidden.

This atStart situation is already looked for when the document is
not protected. However, when the section is already protected, then
a different code path is followed, and there the start position
was ignored and thus became an editable part of the page,
allowing the user to easily delete the field by mistake.

LO6.2's commit 63ec2e8d07dff652a68c9dc96859c4c61f26b9df
accurately identfied that this is in a protected section,
and so exposed this missing scenario.

Related idea: Also added a unit test to ensure that Protect_Form
keeps protecting the formfields as well.
A test document in the bug report, containing a combobox and checkbox,
confirms that this fix is good for all three types of fieldmarks.

Change-Id: Ida08163f3376e55c620d6670da160832e79cd252
Reviewed-on: https://gerrit.libreoffice.org/71927
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>