tdf#125170 Deliver mouse events also when clicking into document content
Event listeners are added once VCLXWindow::SetWindow is called.
This never happened for the document content window.
So we need to call Window::GetComponentInterface which will create an XWindowPeer
and then call UnoWrapper::SetWindowInterface which calls VCLXWindow::SetWindow.

After that, event listeners are registered so that we can deliver events.

Change-Id: I2ed01f1ab20d87fedc4803bdbd96f5ef14d40043
Reviewed-on: https://gerrit.libreoffice.org/71948
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>