vcl: LOK: emmit size_changed after window creation
This guarantees that the correct size of the window
is set on the client side. Some windows are created
with default (and invalid) size and rely on Resize
to be called right after. This doesn't happen
with message boxes and they need to get size_changed
right after the window is fully created and ready.

Change-Id: I8c281a31770de4b37b9de6d333159e5379344dff
Reviewed-on: https://gerrit.libreoffice.org/76293
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>