vcl: Add ScanlineTransformer and make tests
This is intended to replace the ImplPixelFormat in macOS and
OpenGL backends, which are duplicated. It'll also come in
handy for other situations in the future.

Also ImplPixelFormat4 (4-bit Palette) in the macOS backend
doesn't work and never has worked correctly (bug at writing).
This fact was discovered while testing the ScanlineTranformer,
which was based on the ImplPixelFormat* code.

Change-Id: I96a172bd61c2cc116ab3e8b9f47f23f945243e96
Reviewed-on: https://gerrit.libreoffice.org/71992
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>