tdf#90297 use field colors for math's edit engine
This adds a maTextColor to the EditEngine, which will override
the default AutoColor.

There is also a NOCOLORS mode and I thought about using it
instead, but it defaults to COL_BLACK and overriding this seems
wrong.

Since the EditEngine does the whole drawing and now has the
explicit FieldColor() and FieldTextColor() for background and
text, my guess is we don't need an ApplySettings implementation.

Change-Id: I6c2ecc843a3e07de98f97c86fc00a3b83cb78d97
Reviewed-on: https://gerrit.libreoffice.org/72197
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>