VCL keep / return the original set TextEngine font
TextEngine::SetFont copies the font parameter and then modifies
its color, transparency, fill color and alignment, so passing the
same font again will not be detected as the same font.

Therefore this patch stores the original font in addition to the
modified one and also returns that one on the GetFont call.

This allows us to merge the font setup in VclMultiLineEdit's
ImplInitSettings and ApplySettings into a common function.

Change-Id: I618d283ecd0ae14faf9b87a3aceec6ca1c37b526
Reviewed-on: https://gerrit.libreoffice.org/72788
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>