Fix moving / orientation-changing ScrollBar
Just implements Move the same way then Resize. More importantly
the patch correctly resets the Thumb and Page rectangles to
position (0,0) instead of just Empty, which ImplCalc is based on.

Otherwise this results in broken scroll bars, when the StarMath
elements window is docked in in the bottom or top area and
switches the scrolling orientation on undock.

Change-Id: I32b0507cdd6551cc7f42655a730faf8ef25b747b
Reviewed-on: https://gerrit.libreoffice.org/72794
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>