tdf#125319 DOCX track changes: don't change numbering
of a paragraph after a tracked deletion in Hide Changes
mode (including hiding the numbering completely).

We add the numbering of the paragraph to the first node
of the deletion, so Hide Changes mode will show the
actual text content correctly.

Note: file saving already makes the same changes in the
deletion without this workaround, too.

Change-Id: I09d3ad2c05bfd6c7f9f65cfe21a4bade2ac18a14
Reviewed-on: https://gerrit.libreoffice.org/72784
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>